% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairRegression-package.R
\docType{package}
\name{fairRegression-package}
\alias{fairRegression}
\alias{fairRegression-package}
\title{fairRegression}
\description{
Tools for evaluating fairness in regression models.
}
\author{
\strong{Maintainer}: Lin Li \email{lilylovelin37@gmail.com}

}
\keyword{internal}
